feat: implement recommendations UI for web and android, and resolve build errors

+++ web-film-android/app/src/main/java/com/tamdao/cinestream/data/repository/MovieRepository.kt
@@ -438,6 +438,6 @@ class MovieRepository @Inject constructor(
         quality = "HD",
         currentEpisode = null,
         language = null,
-        viewCount = views ?: 0
+        viewCount = (views ?: 0).toLong()
     )
 }
